Make sure we get oncalls for the daytime
We'll more than likely be running this script in the wee small hours, which
means `.on_calls` will return the out of hours people, rather than those who
are covering first line in the day. We can get around this by asking for the
oncalls at midday on that particular day.