Only make changes to group where necessary
When someone is removed from a group, all tickets assigned to them as
part of that group membership are unassigned from the individual and
reassigned to the group. Whilst this is intended behaviour when someone
leaves their support rota, we should not be unassigning people mid-rota.

This commit introduces checks that the action of either removing or
adding a person from the support group is necessary to make the group
match the rota.